+++ handlers/login.go
@@ -55,7 +55,7 @@ func LoginHandler(w http.ResponseWriter, r *http.Request) {
 
 	// set the path for the session cookie to only send the correct cookie to /auth/{state}/
 	// must have a trailing slash. Otherwise, it is send to all endpoints that _start_ with the cookie path.
-	session.Options.Path = fmt.Sprintf("/auth/%s/", state)
+	session.Options.Path = fmt.Sprintf("%s/auth/%s/", cfg.Cfg.DocumentRoot, state)
 
 	log.Debugf("session state set to %s", session.Values["state"])